Fix cascading two modifiers.
We can't really process one morphed code then pass it to the next morphing.
This will falsify the location/range of the original function as it may be
shifted by the extra instrumentation.

The solution is to collect first all the code fragments to be inserted and
then apply them at once.